Updated and rename
The name of the .md file was changed to reflect the content so it's easier to find. The name was "contributing_in_agile_org.md" and the title was "How to contribute to ThreeFold". 
Under the requirements for a TF Founder, there is a list of docs that need to be read. I have added on top the FF Manifesto. 
Please clarify if there any other docs that need to be added here and if the Efika culture manifest and other docs under dividi / values still apply or might need some revision.